1 итерация

+++ src/components/card.js
@@ -1,33 +1,60 @@
+import {createLikeCard, deleteLikeCard} from './api.js';
+
 const cardTemplate = document.querySelector('#card-template').content;
 
 export function createCard(cardData, onDeleteClick, onLikeClick, onImageClick) {
   const cardElement = cardTemplate.querySelector('.card').cloneNode(true);
+
   const cardImage = cardElement.querySelector('.card__image');
   cardImage.src = cardData.link;
   cardImage.alt = cardData.name;
   cardElement.querySelector('.card__title').textContent = cardData.name;
 
+  cardElement.dataset.cardId = cardData._id;
+
   const deleteButton = cardElement.querySelector('.card__delete-button');
-  deleteButton.addEventListener('click', () => onDeleteClick(cardElement));
+
+  if (onDeleteClick) {
+    deleteButton.addEventListener('click', () => onDeleteClick(cardElement));
+  } else {
+    deleteButton.style.display = 'none';
+  };
 
   const likeButton = cardElement.querySelector('.card__like-button');
-  likeButton.addEventListener('click', onLikeClick);
+  likeButton.addEventListener('click', (evt) => onLikeClick(evt, cardElement));
 
   const cardLikeCount = cardElement.querySelector('.card__like-count');
   cardLikeCount.textContent = cardData.likes.length; 
 
   cardImage.addEventListener('click', () => onImageClick(cardData.link, cardData.name));
 
- 
   return cardElement;
 };
 
-export function handleLikeClick(evt) {
-  evt.target.classList.toggle('card__like-button_is-active');
+export function handleLikeClick(evt, cardElement) {
+  const liked = evt.target.classList.toggle('card__like-button_is-active');
+
+  const cardLikeCount = cardElement.querySelector('.card__like-count');
+  const cardId = cardElement.dataset.cardId;
+
+  if(liked) {
+    createLikeCard(cardId)
+      .then((data) => {
+        cardLikeCount.textContent = data.likes.length;
+      })
+
+      .catch((err) => {
+        console.log('There was a problem with the update operation:', err);
+      });
+
+  } else {
+    deleteLikeCard(cardId)
+      .then((data) => {
+        cardLikeCount.textContent = data.likes.length;
+      })
+
+      .catch((err) => {
+        console.log('There was a problem with the update operation:', err);
+      });
+    }
 };
-
-export function deleteClick(card) {
-  card.remove();
-};
-
-
